+++ src/g_text.cpp
@@ -104,7 +104,7 @@ public:
     ~GFramePeer();
     
     int ConSetTitle(char *Title, char *STitle);
-    int ConGetTitle(char *Title, int MaxLen, char *STitle, int SMaxLen);
+    int ConGetTitle(char *Title, size_t MaxLen, char *STitle, size_t SMaxLen);
     
     int ConSetSize(int X, int Y);
     int ConQuerySize(int *X, int *Y);
@@ -573,7 +573,7 @@ int GFramePeer::ConSetTitle(char *Title, char *STitle) {
     return 0;
 }
 
-int GFramePeer::ConGetTitle(char *Title, int MaxLen, char *STitle, int SMaxLen) {
+int GFramePeer::ConGetTitle(char *Title, size_t MaxLen, char *STitle, size_t SMaxLen) {
     return ::ConGetTitle(Title, MaxLen, STitle, SMaxLen);
 }
 
@@ -616,7 +616,7 @@ int GFrame::ConSetTitle(char *Title, char *STitle) {
     return Peer->ConSetTitle(Title, STitle);
 }
 
-int GFrame::ConGetTitle(char *Title, int MaxLen, char *STitle, int SMaxLen) {
+int GFrame::ConGetTitle(char *Title, size_t MaxLen, char *STitle, size_t SMaxLen) {
     return Peer->ConGetTitle(Title, MaxLen, STitle, SMaxLen);
 }
